# Conflicts:
#	src/main/java/com/controller/MainController.java
#	src/main/resources/properties.properties

+++ src/main/java/com/repository/statistic/StaticRepo.java
@@ -9,7 +9,7 @@ import java.util.Date;
 import java.util.List;
 
 @Repository
-public interface StaticRepo extends JpaRepository<Static, Long> {
+public interface StaticRepo extends JpaRepository<Static, Integer> {
     @Query("select s from Static s where s.date = :date")
     List<Static> getAllByDate(Date date);
 }
